bug(loop extract)

+++ src/Compiler.java
@@ -55,12 +55,9 @@ public class Compiler {
         dataFlow.divideBlock();
         dataFlow.arriveDataAnalysis();
         dataFlow.activeDataAnalysis();
-        if (optimize.get("DeleteDeadCode")) {
-            dataFlow.deleteDeadCode();
-        }
-        if (optimize.get("BroadcastCode")) {
-            dataFlow.broadcastCode();
-        }
+        if (optimize.get("DeleteDeadCode")) dataFlow.deleteDeadCode();
+        if (optimize.get("BroadcastCode")) dataFlow.broadcastCode();
+        if (optimize.get("ExtractLoopConstExp")) dataFlow.extractLoopConstExp();
         MidCodeList.codes = dataFlow.getCodes();
         write(MidCodeList.printMidCode(), "ir_opt.txt");
         // dataFlow.printGraph();
